refactor: remove python wrapper (#3859)
the python wrapper exists because
we should implement __setstate__ and __getstate__for
ECDHCurve25519 so it can use serialization and deserialization
from pickle. We find way to implement these methods in rust.
